+++ lib/carto/connector/connection_adapter/bigquery.rb
@@ -40,21 +40,21 @@ module Carto
 
       def create
         super
-        return if @connection.parameters.blank?
+        return unless spatial_extension?
 
         create_spatial_extension_setup
       end
 
       def destroy
         super
-        return if @connection.parameters.blank?
+        return unless spatial_extension?
 
         remove_spatial_extension_setup
       end
 
       def update
         super
-        return if @connection.parameters.blank?
+        return unless spatial_extension?
 
         update_spatial_extension_setup
       end
@@ -66,6 +66,10 @@ module Carto
       # end
 
       private
+
+      def spatial_extension?
+        @connection.parameters.present? && !@connection.shared?
+      end
 
       def central
         @central ||= Cartodb::Central.new
@@ -117,30 +121,6 @@ module Carto
         { 'token': @connection.token }
       end
 
-      # Temporally we need to mantain this redis key until maps api is updated
-      def bigquery_redis_key
-        "google:bq_settings:#{@connection.user.username}"
-      end
-
-      def update_redis_metadata
-        super
-
-        return unless @connection.parameters.present?
-
-        if @connection.parameters['service_account'].present?
-          $users_metadata.hmset(
-            bigquery_redis_key,
-            'service_account', @connection.parameters['service_account'],
-            'billing_project', @connection.parameters['billing_project']
-          )
-        end
-      end
-
-      def remove_redis_metadata
-        super
-
-        $users_metadata.del bigquery_redis_key
-      end
     end
   end
 end